[ALPS06569827] cam: kernel side
[Detail]
1.  fix conflict
3-4.remove debug log
5.  real sensor support tg 1 pixel
6.  fix coverity

MTK-Commit-Id: 2faa02365cb3a780381d88891647c49d9791d496

CR-Id: ALPS06569827
Feature: [Android Default] Camera Application Basic Functions

+++ drivers/misc/mediatek/imgsensor/src-v4l2/adaptor-drv.c
@@ -1113,7 +1113,7 @@ static ssize_t debug_pwr_ops_show(struct device *dev,
 		else
 			ent = &ctx->subdrv->pw_seq[i];
 		SHOW(buf, len, "\t%s power %d with delay %d\n",
-		     hw_id_names[ent->id], ent->val, ent->delay);
+		     hw_id_names[(unsigned int)ent->id], ent->val, ent->delay);
 	}
 
 	return len;
